requirements.txt
streamlit>=1.22.0
crewai>=0.28.0
openai>=1.3.0
langchain>=0.0.271
langgraph>=0.0.16
ibm-watsonx-ai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0